ASAP-183 fix: 행성 내의 편지 수 반환하지 않는 문제 해결

+++ Application-Module/src/main/kotlin/com/asap/application/space/port/in/SpaceGetUsecase.kt
@@ -20,7 +20,7 @@ interface SpaceGetUsecase {
 
     data class SpaceDetail(
         val spaceName: String,
-        val letterCount: Int,
+        val letterCount: Long,
         val isMainSpace: Boolean,
         val spaceIndex: Int,
         val spaceId: String,
